src/requirements.txt
asyncpg==0.23.0
autopep8==1.5.7
bcrypt==3.2.0
cffi==1.14.5
click==7.1.2
cryptography==3.4.7
dnspython==2.1.0
ecdsa==0.14.1
email-validator==1.1.2
fastapi==0.65.1
greenlet==1.1.0
h11==0.12.0
idna==3.2
passlib==1.7.4
psycopg2-binary==2.8.6
pyasn1==0.4.8
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.2
PyJWT==2.1.0
python-jose==3.2.0
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.17
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.0
uuid==1.30
uvicorn==0.13.4
